// src/components/UI-kit/BaseControls/inputs/InputPhoneUI/InputPhoneUI.types.ts
import {InputHTMLAttributes} from 'react'
type TTheme = 'white' | 'dark'
export interface IField extends InputHTMLAttributes<HTMLInputElement> {
  placeholder?: string
  error?: string
  extraClass?: string
  type?: string
  extraStyle?: React.CSSProperties
  icon?: React.ReactNode
  labelText?: string
  extraClassClearIcon?: string
  disabled?: boolean
  theme?: TTheme
  customPattern?: RegExp
}